Re-arranged keyset and en-keyset modules. Turned Keyset into Interface
